Partial fix for Manual Form closing problem
When using Form.ShowDialog and invoking Form.Close, sometimes the dialog
cannot be closed for some reasons. Such issue happens in both Client and
Orchestrator. This patch fixes totally for the Orchestrator and
partially for Client.
The problem right now is that after several manual operations, the
Client stops listening to the Orchestrator.